requirements.txt
apscheduler
hikari
hikari-lightbulb